Solved #10 by not displaying the drone when overlapping an obstacle
There is no way that the drone won't go on an obstacle cell. It's gonna
happen because the obstacles are a point and even if a very strong repulsive
force is applied if the drone is fast enough it will go over it. One other
solution is set the drone speed very slow but I don't like the idea too much
or raise again the force of the obstacles but then becomes annoying